Add alert dialog to show user progress
